Several fixes & improvements

+++ lib/src/utils/file_utils.dart
@@ -44,12 +44,7 @@ Future<File> createArbFileForLocale(String locale, String arbDir) async {
 /// Gets all arb files in the project.
 List<FileSystemEntity> getArbFiles(String arbDir) {
   var l10nDirPath = path.join(getRootDirectoryPath(), arbDir);
-  var arbFiles = Directory(l10nDirPath)
-      .listSync()
-      .where((file) =>
-          path.basename(file.path).startsWith('intl_') &&
-          path.basename(file.path).endsWith('.arb'))
-      .toList();
+  var arbFiles = Directory(l10nDirPath).listSync().where((file) => path.basename(file.path).startsWith('intl_') && path.basename(file.path).endsWith('.arb')).toList();
 
   // arb files order is not the same on all operating systems (e.g. win, mac)
   arbFiles.sort((a, b) => a.path.compareTo(b.path));
@@ -59,18 +54,13 @@ List<FileSystemEntity> getArbFiles(String arbDir) {
 
 /// Gets all locales in the project.
 List<String> getLocales(String arbDir) {
-  var locales = getArbFiles(arbDir)
-      .map((file) => path.basename(file.path))
-      .map((fileName) =>
-          fileName.substring('intl_'.length, fileName.length - '.arb'.length))
-      .toList();
+  var locales = getArbFiles(arbDir).map((file) => path.basename(file.path)).map((fileName) => fileName.substring('intl_'.length, fileName.length - '.arb'.length)).toList();
 
   return locales;
 }
 
 /// Updates arb file content.
-Future<void> updateArbFile(
-    String fileName, Uint8List bytes, String arbDir) async {
+Future<void> updateArbFile(String fileName, Uint8List bytes, String arbDir) async {
   var rootDirPath = getRootDirectoryPath();
   var arbFilePath = path.join(rootDirPath, arbDir, fileName);
   var arbFile = File(arbFilePath);
@@ -83,8 +73,7 @@ Future<void> updateArbFile(
 }
 
 /// Gets l10n Dart file path.
-String getL10nDartFilePath(String outputDir) =>
-    path.join(getRootDirectoryPath(), outputDir, 'l10n.dart');
+String getL10nDartFilePath(String outputDir) => path.join(getRootDirectoryPath(), outputDir, 'l10n.dart');
 
 /// Updates l10n Dart file.
 Future<void> updateL10nDartFile(String content, String outputDir) async {
@@ -99,8 +88,7 @@ Future<void> updateL10nDartFile(String content, String outputDir) async {
 }
 
 /// Gets intl directory path.
-String getIntlDirectoryPath(String outputDir) =>
-    path.join(getRootDirectoryPath(), outputDir, 'intl');
+String getIntlDirectoryPath(String outputDir) => path.join(getRootDirectoryPath(), outputDir);
 
 /// Gets intl directory.
 Directory? getIntlDirectory(String outputDir) {
@@ -123,8 +111,7 @@ Future<Directory> createIntlDirectory(String outputDir) async {
 }
 
 /// Removes unused generated Dart files.
-Future<void> removeUnusedGeneratedDartFiles(
-    List<String> locales, String outputDir) async {
+Future<void> removeUnusedGeneratedDartFiles(List<String> locales, String outputDir) async {
   var intlDir = getIntlDirectory(outputDir);
   if (intlDir == null) {
     return;
@@ -133,13 +120,12 @@ Future<void> removeUnusedGeneratedDartFiles(
   var files = intlDir.listSync();
   for (var file in files) {
     var basename = path.basename(file.path);
-    var substring = basename.substring(
-        'messages_'.length, basename.length - '.dart'.length);
+    if (basename.contains('messages')) {
+      var substring = basename.substring('messages_'.length, basename.length - '.dart'.length);
 
-    if (basename.startsWith('messages_') &&
-        basename.endsWith('.dart') &&
-        !['all', ...locales].contains(substring)) {
-      await file.delete(recursive: true);
+      if (basename.startsWith('messages_') && basename.endsWith('.dart') && !['all', ...locales].contains(substring)) {
+        await file.delete(recursive: true);
+      }
     }
   }
 }
